requirements.txt
absl-py==1.2.0
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-appconf==1.0.5
django-braces==1.15.0
django-cors-headers==3.13.0
django-imagekit==2.0.4
django-summernote==0.8.20.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
executing==1.1.1
fonttools==4.37.4
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
idna==3.4
ipython==8.5.0
jedi==0.18.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.2
seaborn==0.12.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.4.0
types-cryptography==3.3.23
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
